chore: move test together to the tested code
